requirements.txt
openai
speechrecognition
gtts
flask
python-dotenv
pyaudio